requirements.txt
langgraph>=0.0.40
requests>=2.31.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

